Move the afffix code into it's own directory
