+++ link-grammar/dict-common/dict-common.h
@@ -16,80 +16,14 @@
 
 #include "dict-structures.h"
 
-/* The functions here are for link-grammar internal use only.
- * They are not part of the public API. */
+/* The functions here are intended for use by the tokenizer, only,
+ * and pretty much no one else. If you are not the tokenizer, you
+ * probably dont need these. */
 
 bool find_word_in_dict(Dictionary dict, const char *);
-Afdict_class * afdict_find(Dictionary, const char *, bool);
 
 Exp * Exp_create(Exp_list *);
 void add_empty_word(Dictionary const, X_node *);
 void free_Exp_list(Exp_list *);
-
-void patch_subscript(char *);
-
-bool is_suffix(const char, const char *);
-bool is_stem(const char *);
-
-/* Connector names for the affix class lists in the affix file */
-
-typedef enum {
-	AFDICT_RPUNC=1,
-	AFDICT_LPUNC,
-	AFDICT_UNITS,
-	AFDICT_SUF,
-	AFDICT_PRE,
-	AFDICT_MPRE,
-	AFDICT_QUOTES,
-	AFDICT_BULLETS,
-	AFDICT_INFIXMARK,
-	AFDICT_STEMSUBSCR,
-	AFDICT_SANEMORPHISM,
-
-	/* The below are used only for random morphology via regex */
-	AFDICT_REGPRE,
-	AFDICT_REGMID,
-	AFDICT_REGSUF,
-	AFDICT_REGALTS,
-	AFDICT_REGPARTS,
-
-	/* Have to have one last entry, to get the array size corrrect */
-	AFDICT_LAST_ENTRY,
-	AFDICT_NUM_ENTRIES
-} afdict_classnum;
-
-#define AFDICT_CLASSNAMES1 \
-	"invalid classname", \
-	"RPUNC", \
-	"LPUNC", \
-	"UNITS", \
-	"SUF",         /* SUF is used in the Russian dict */ \
-	"PRE",         /* PRE is not used anywhere, yet... */ \
-	"MPRE",        /* Multi-prefix, currently for Hebrew */ \
-	"QUOTES", \
-	"BULLETS", \
-	"INFIXMARK",   /* Prepended to suffixes, appended to pefixes */ \
-	"STEMSUBSCR",  /* Subscripts for stems */ \
-	"SANEMORPHISM", /* Regex for sane_morphism() */
-
-/* The regexes below are used only for random morphology generation */
-#define AFDICT_CLASSNAMES2 \
-	"REGPRE",      /* Regex for prefix */ \
-	"REGMID",      /* Regex for middle parts */ \
-	"REGSUF",      /* Regex for suffix */ \
-	"REGALTS",     /* Min&max number of alternatives to issue for a word */\
-	"REGPARTS",    /* Max number of word partitions */
-
-#define AFDICT_CLASSNAMES AFDICT_CLASSNAMES1 AFDICT_CLASSNAMES2 "last classname"
-#define AFCLASS(afdict, class) (&afdict->afdict_class[class])
-
-/* Suffixes start with it.
- * This is needed to distinguish suffixes that were stripped off from
- * ordinary words that just happen to be the same as the suffix.
- * Kind-of a weird hack, but I'm not sure what else to do...
- * Similarly, prefixes end with it.
- */
-#define INFIX_MARK(afdict) \
-	((NULL == afdict) ? '\0' : (AFCLASS(afdict, AFDICT_INFIXMARK)->string[0][0]))
 
 #endif /* _LG_DICT_COMMON_H_ */
