+++ link-grammar/dict-file/word-file.c
@@ -18,6 +18,27 @@
 #include "read-dict.h"
 #include "word-file.h"
 
+/** Replace the right-most dot with SUBSCRIPT_MARK */
+void patch_subscript(char * s)
+{
+	char *ds, *de;
+	int dp;
+	ds = strrchr(s, SUBSCRIPT_DOT);
+	if (!ds) return;
+
+	/* a dot at the end or a dot followed by a number is NOT
+	 * considered a subscript */
+	de = ds + 1;
+	if (*de == '\0') return;
+	dp = (int) *de;
+
+	/* If its followed by a UTF8 char, its NOT a subscript */
+	if (127 < dp || dp < 0) return;
+	/* assert ((0 < dp) && (dp <= 127), "Bad dictionary entry!"); */
+	if (isdigit(dp)) return;
+	*ds = SUBSCRIPT_MARK;
+}
+
 /**
  * Reads in one word from the file, allocates space for it,
  * and returns it.
